move inject tasks to file
