+++ generators/client-2/templates/gulp/config.js
@@ -1,3 +1,5 @@
+'use strict';
+
 module.exports = {
     app: '<%= MAIN_SRC_DIR %>',
     dist: '<%= DIST_DIR %>',
