+++ generators/client-2/templates/gulp/serve.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var gulp = require('gulp'),
     util = require('./utils'),
     url = require('url'),
